Separated createProject to testable fragments
