+++ packages/core/cli/src/scripts/createProject/buildFileTargetPath.ts
@@ -1,0 +1,1 @@
+module.exports = (file: string, targetPath: string, chopPhrase: string): string => targetPath + (file.replace(chopPhrase, ''));
