+++ packages/core/cli/src/scripts/createProject/getThemePath.ts
@@ -1,0 +1,1 @@
+module.exports = (themeName: string): string => `../../../node_modules/@vue-storefront/${themeName}`;
