+++ packages/core/cli/src/scripts/createProject/removeMagicCommentsFromFile.ts
@@ -1,0 +1,11 @@
+const fs = require('fs');
+
+module.exports = async (absoluteFilePath: string): Promise<void> => {
+  const removeDevMagicComment = (source: string): string => source.replace(/\s+(\/\/ @core-development-only-start)(.*?)(\/\/ @core-development-only-end)/sg, '');
+  const fileContent = fs.readFileSync(absoluteFilePath, { encoding: 'utf8' });
+
+  return fs.writeFileSync(
+    absoluteFilePath,
+    removeDevMagicComment(fileContent)
+  );
+};
